+++ src/main/java/ru/sbt/mipt/oop/HomeBuilder.java
@@ -2,6 +2,8 @@ package ru.sbt.mipt.oop;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import ru.sbt.mipt.oop.base_device.Door;
+import ru.sbt.mipt.oop.base_device.Light;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
