+++ src/main/java/ru/sbt/mipt/oop/RandomEventGenerator.java
@@ -1,6 +1,6 @@
 package ru.sbt.mipt.oop;
 
-public class EventGenerator {
+public class RandomEventGenerator {
 
     public static SensorEvent getNextSensorEvent() {
         // pretend like we're getting the events from physical world, but here we're going to just generate some random events
