+++ src/main/java/ru/sbt/mipt/oop/Room.java
@@ -1,11 +1,14 @@
 package ru.sbt.mipt.oop;
 
+import ru.sbt.mipt.oop.base_device.Door;
+import ru.sbt.mipt.oop.base_device.Light;
+
 import java.util.Collection;
 
 public class Room {
-    private Collection<Light> lights;
-    private Collection<Door> doors;
-    private String name;
+    private final Collection<Light> lights;
+    private final Collection<Door> doors;
+    private final String name;
 
     public Room(Collection<Light> lights, Collection<Door> doors, String name) {
         this.lights = lights;
