+++ src/main/java/ru/sbt/mipt/oop/base_device/Door.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.base_device;
 
 public class Door {
     private final String id;
