+++ src/main/java/ru/sbt/mipt/oop/reader/JsonSmartHomeReader.java
@@ -1,6 +1,8 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.reader;
 
 import com.google.gson.Gson;
+import ru.sbt.mipt.oop.SmartHome;
+import ru.sbt.mipt.oop.reader.SmartHomeReader;
 
 import java.io.IOException;
 import java.nio.file.Files;
